+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
     "airflow-operator",
     "commons",
     #"druid-operator",
-    "hive-operator",
     #"kafka-operator",
     # "monitoring-operator",
     # "nifi-operator",
