--- hive-operator/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-authors = ["Stackable GmbH <info@stackable.de>"]
-description = "Integration tests for the Stackable Operator for Apache Hive"
-edition = "2021"
-license = "OSL-3.0"
-name = "hive-operator-integration-tests"
-repository = "https://github.com/stackabletech/integration-tests"
-version = "0.1.0"
-
-[dependencies]
-anyhow = "1.0"
-integration-test-commons = { path = "../commons" }
-stackable-hive-crd = { git = "https://github.com/stackabletech/hive-operator.git", branch = "main" }
-serde = "1.0"
-serde_json = "1.0"
-serde_yaml = "0.8"
