+++ hive-operator/tests/kuttl/smoke/00-assert.yaml
@@ -1,0 +1,14 @@
+---
+apiVersion: kuttl.dev/v1beta1
+kind: TestAssert
+metadata:
+  name: install-test-hive-derby
+timeout: 600
+---
+apiVersion: apps/v1
+kind: StatefulSet
+metadata:
+  name: test-hive-derby-metastore-default
+status:
+  readyReplicas: 1
+  replicas: 1
