+++ hive-operator/tests/kuttl/smoke/01-assert.yaml
@@ -1,0 +1,13 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestAssert
+metadata:
+  name: test-metastore
+timeout: 300
+---
+apiVersion: apps/v1
+kind: StatefulSet
+metadata:
+  name: test-metastore
+status:
+  readyReplicas: 1
+  replicas: 1
