+++ hive-operator/tests/kuttl/smoke/02-prepare-test-metastore.yaml
@@ -1,0 +1,6 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+commands:
+  - script: kubectl cp -n $NAMESPACE ./test_metastore.py  test-metastore-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./requirements.txt test-metastore-0:/tmp
+  - script: kubectl exec -n $NAMESPACE test-metastore-0 -- pip install --user -r /tmp/requirements.txt
