+++ hive-operator/tests/kuttl/smoke/requirements.txt
